[QLN-108] ref: refactor code
