--- src/app/[locale]/(main)/classrooms/[id]/manage/layout.tsx
@@ -1,20 +1,0 @@
-import BackgroundSquare from "@/components/background-square"
-import { getUser } from "@/lib/auth"
-import { notFound } from "next/navigation"
-
-type Props = { children?: React.ReactNode }
-
-function ManageClassroomLayout({ children }: Props) {
-  const user = getUser()
-  if (!user) return notFound()
-  return (
-    <BackgroundSquare
-      variant={"topDown"}
-      className="container mx-auto items-start "
-    >
-      {children}
-    </BackgroundSquare>
-  )
-}
-
-export default ManageClassroomLayout
