+++ src/app/[locale]/(main)/classrooms/[id]/members/components/ban-dialog.tsx
