+++ src/app/[locale]/(main)/classrooms/[id]/members/components/student-table.tsx
@@ -27,7 +27,7 @@ import {
   getSortedRowModel,
   useReactTable,
 } from "@tanstack/react-table"
-import { useTranslations } from "next-intl"
+import { useFormatter, useTranslations } from "next-intl"
 import React from "react"
 import BanDialog from "./ban-dialog"
 
@@ -42,6 +42,7 @@ const StudentTable = ({ data }: StudentTableProps) => {
   )
   const t = useTranslations("Table")
   const i18n = useTranslations("Classroom_student")
+  const format = useFormatter()
 
   const columns: ColumnDef<User>[] = React.useMemo(
     () => [
@@ -86,15 +87,21 @@ const StudentTable = ({ data }: StudentTableProps) => {
         },
       },
       {
-        header: i18n("table.headers.role"),
-        accessorKey: "role",
+        header: i18n("table.headers.dob"),
+        accessorKey: "dob",
         cell: ({ row }) => {
-          const role = row.getValue("role") as User["role"]
-          return <div className="min-w-52 capitalize">{role}</div>
+          const dob = row.getValue("dob") as User["dob"]
+          return (
+            <div className="min-w-52">
+              {format.dateTime(new Date(dob), {
+                dateStyle: "long",
+              })}
+            </div>
+          )
         },
       },
     ],
-    [i18n]
+    [format, i18n]
   )
 
   const [sorting, setSorting] = React.useState<SortingState>([])
